+++ packages/core/lib/core.dart
@@ -1,6 +1,9 @@
 /// Multichoice Core
 library core;
 
-export 'src/application/export_application.dart';
+export 'src/application/export.dart';
 export 'src/get_it_injection.dart';
-export 'src/services/export_services.dart';
+export 'src/repositories/export.dart';
+export 'src/services/export.dart';
+export 'src/utils/export.dart';
+export 'src/wrappers/export.dart';
