+++ packages/core/lib/src/repositories/export.dart
@@ -1,0 +1,2 @@
+export 'interfaces/entry/i_entry_repository.dart';
+export 'interfaces/tabs/i_tabs_repository.dart';
