--- packages/core/lib/src/repositories/export_repositories.dart
@@ -1,2 +1,0 @@
-export 'implementation/entry/entry_repository.dart';
-export 'implementation/tabs/tabs_repository.dart';
