+++ packages/core/lib/src/services/export.dart
@@ -1,3 +1,2 @@
 export 'interfaces/i_app_info_service.dart';
-export 'interfaces/i_database_service.dart';
 export 'interfaces/i_data_exchange_service.dart';
